move updateRole method from AuthService to RoleService
