+++ src/main/java/com/vu/aezakmi/service/AuthService.java
@@ -48,21 +48,5 @@ public class AuthService {
         return new UserDto(user);
     }
 
-    public void updateRole(Long userId, RoleType roleType) {
-        Role adminRole = roleRepository.findByType(RoleType.ADMIN)
-                .orElseThrow(() -> new RuntimeException("Admin role not found"));
 
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new RuntimeException("User not found"));
-
-        if (!user.getRole().equals(adminRole)) {
-            throw new RuntimeException("Only admin can update user roles");
-        }
-
-        Role role = roleRepository.findByType(roleType)
-                .orElseThrow(() -> new RuntimeException("Role not found"));
-
-        user.setRole(role);
-        userRepository.save(user);
-    }
 }
